refactor(oauth): restructure OAuth API endpoints and enhance request handling
