+++ guillotina/contrib/oauth/storage/utility.py
@@ -14,7 +14,6 @@ from guillotina.interfaces import IApplication, IDatabase
 
 logger = logging.getLogger("guillotina.contrib.oauth")
 
-_ddl_locks = {}
 _ddl_initialized = set()
 
 OAUTH_STORAGE_DEFAULTS = {
@@ -43,23 +42,20 @@ def get_oauth_storage_settings():
 async def ensure_oauth_tables(storage):
     import asyncpg.exceptions
 
-    loop = asyncio.get_running_loop()
-    lock = _ddl_locks.setdefault(id(loop), asyncio.Lock())
     storage_key = id(storage.pool)
-    async with lock:
-        if storage_key in _ddl_initialized:
-            return
-        async with storage.pool.acquire() as conn:
-            for ddl in OAUTH_DDL:
-                for attempt in range(3):
-                    try:
-                        await conn.execute(ddl)
-                        break
-                    except asyncpg.exceptions.UniqueViolationError:
-                        if attempt == 2:
-                            raise
-                        await asyncio.sleep(0.05)
-        _ddl_initialized.add(storage_key)
+    if storage_key in _ddl_initialized:
+        return
+    async with storage.pool.acquire() as conn:
+        for ddl in OAUTH_DDL:
+            for attempt in range(3):
+                try:
+                    await conn.execute(ddl)
+                    break
+                except asyncpg.exceptions.UniqueViolationError:
+                    if attempt == 2:
+                        raise
+                    await asyncio.sleep(0.05)
+    _ddl_initialized.add(storage_key)
 
 
 @implementer(IOAuthStorageUtility)
